+++ src/org/proj/math/tensor/Tensor3D.java
@@ -3,6 +3,7 @@ package org.proj.math.tensor;
 import org.proj.math.matrix.Matrix;
 import org.proj.math.vector.Vector;
 
+import java.math.BigDecimal;
 import java.util.Iterator;
 
 public abstract class Tensor3D implements Iterable<Matrix> {
@@ -14,12 +15,12 @@ public abstract class Tensor3D implements Iterable<Matrix> {
         this.gamma = gamma;
     }
 
-    public abstract double get (int x, int y, int z);
+    public abstract BigDecimal get (int x, int y, int z);
 
     public Matrix get (int i) {
         return new Matrix (beta, gamma) {
             @Override
-            public double get (int j, int k) {
+            public BigDecimal get (int j, int k) {
                 return Tensor3D.this.get(i, j, k);
             }
         };
@@ -28,7 +29,7 @@ public abstract class Tensor3D implements Iterable<Matrix> {
     public Vector get (int i, int j) {
         return new Vector (gamma) {
             @Override
-            public double get(int k) {
+            public BigDecimal get(int k) {
                 return Tensor3D.this.get(i, j, k);
             }
         };
@@ -71,7 +72,7 @@ public abstract class Tensor3D implements Iterable<Matrix> {
         public abstract Matrix get (int i);
 
         @Override
-        public double get (int x, int y, int z) {
+        public BigDecimal get (int x, int y, int z) {
             return get(x).get(y, z);
         }
     }
